Hypsographic Functions from Tabular Interpolation

The main function creates a function list object from a data table to allow interpolation of depths, areas and volume fractions of layers.

Usage

hypso_functions(table)

Arguments

table

hypsographic table (bathymetry) given as data frame with columns level, area and volume

Value

a list of functions with the following elements (slots):

level(volume) interpolates level from volume

volume(level) interpolates volume from a given water level

area(level) interpolates area from a given level

vh(level, zmixreal) calculates hypolimnic volume from surface water level (m above ground) and mixing depth (zmixreal, thermocline in m below surface)

ve(level, zmixreal) calculates epilimnic volume from surface water level and zmixreal

zmix(level, zmixreal) calculates average mixing depth (zmix) from surface water level and zmixreal

zhm(level, zmixreal) calculates average depth of bottom layer (hypolimnion and metalimnion)

sediment_area(level) calculates sediment area for a series of water levels, where level shoudl be a vecor of water levels in ascending order

pelagic_ratio(level) calculates the area ratio of the pelagic part (water) to the total area of layers. It is equal to 1 - sediment area fraction.

See also

Examples

data(hypso_bautzen)                      # load data table
hypso <- hypso_functions(hypso_bautzen)  # create function object

hypso$volume(166)                        # volume at 166 m a.s.l.
#> [1] 29463587
hypso$sediment_area(c(160, 166))         # sediment area of hypo- and epilimnion
#> [1] 2722336 1914372